+++ lib/template.js
@@ -23,12 +23,13 @@ module.exports = {
     while (i < filelist.length) {
       list = list + `
       <li>
-        <a href="/page/${filelist[i]}">${filelist[i]}</a>
+        <a href="/topic/${filelist[i]}">${filelist[i]}</a>
       </li>`;
 
       i++;
     }
     list += `</ul>`
+    
     return list;
   }
 }
