+++ package.json
@@ -19,6 +19,7 @@
   "homepage": "https://github.com/dev-heeseok/web-nodejs#readme",
   "dependencies": {
     "body-parser": "^1.20.0",
+    "compression": "^1.7.4",
     "express": "^4.18.1",
     "pm2": "^5.2.0",
     "sanitize-html": "^2.7.0"
